requirements.txt
psutil>=5.9.0
ollama>=0.1.0
pydantic>=2.0.0
click>=8.1.0
rich>=13.0.0
jinja2>=3.1.0
pyyaml>=6.0.0
watchdog>=3.0.0
sqlalchemy>=2.0.0
reportlab>=4.0.0
fastapi>=0.100.0
uvicorn[standard]>=0.22.0
python-multipart>=0.0.6
